cocemfe_sevilla/requirements.txt
django-bootstrap-v5
Pillow==9.4.0
folium==0.12.1
coverage==6.5.0
requests==2.31.0
channels==4.1.0
daphne==4.1.2
django-cors-headers==4
django-q==1.3.9